config: Do not early exit when pruning gitaly process dir
When pruning old directories, we want to prune as many as we
can. We don't want to bail early as soon as there is an issue
with one of the entries. For instance, we will error out
and stop trying to prune if there's an error trying to find
the pid.

We can simplify the code to just log the error and continue
each time.