services: Adapt request-validation tests to verify the complete error
With an earlier version of this patch series to start detecting the
object hash format used by a repository some tests performing validation
of requests with invalid repositories have started to fail because we
now wrapped the resulting error. As a consequence, error codes have
changed for those tests to `Internal`, but there wasn't really any
indicator why that was happening.

Refactor these tests that were failing to properly verify the complete
gRPC error returned by the RPCs. Like this, we have much stronger
guarantees about why an RPC call is failing and have more information in
case the test does indeed start to fail in an unexpected way. While not
required anymore for this patch series, this refactoring is still worth
it on its own to improve the test's quality.