repository: Amend test to only intercept some Git commands
One of our tests for the `GetArchive()` RPC uses an intercepting command
factory to assert that git-archive(1) in fact sees expected environment
variables. This is done by unconditionally intercepting any Git command
spawned and replacing it with a call to env(1), which only works right
now because we only ever spawn any other command in this code path at
the moment. This is about to change though with the introduction of the
object hash format detection.

Refactor the test to only intercept calls to git-archive(1) in order for
the test not to break when we start spawning other Git commands.