gittest: Make CreateRepository handle repo creation without gRPC service
We have two sets of functions to create repositories for testing
purposes:

    - CreateRepository creates the repository by calling the respective
      RPCs of the RepositoryService. This allows us to properly test
      with Praefect as a proxy, which requires the RPC to be called so
      that it can set up the repositories in the database.

    - InitRepo and CloneRepo, which both will create the repository
      without doing an RPC call. These should only be used in contexts
      where we don't have a gRPC service available.

It's confusing at times which of both should be used, and we're not
quite consistent.

Unify all functionality to create repositories into CreateRepository and
add a new option `SkipCreationViaService`. If set, this will behave the
same as InitRepo and CloneRepo, which allows us to consolidate the
functions into a single one. Callers that want to skip creation via the
RepositoryService will thus need to explicitly opt-in to this behaviour
and cannot just use InitRepo respectively CloneRepo without knowing
about the ramifications.