gitaly-hooks: Convert test to use Bash instead of Ruby for hooks
One of our tests for gitaly-hooks asserts that both the command line
arguments and environment seen by the process match our expectations.
This is done by writing some custom hooks, where one of the hooks is
currently written in Ruby.

Refactor the hook to use Bash instead in order to reduce our reliance on
Ruby. While at it, convert the test to use `gittest.InitRepo()` instead
of `gittest.InitRepo()`: we don't depend on any references or objects
anyway, so there is not much of a point to use a completely seeded
repository.