operations: Convert test hooks to use Bash instead of Ruby
As we're progressing with out conversion of Ruby-based code we're
drawing ever closer to a future where there is no Ruby in Gitaly's code
base anymore. Some of our tests still depend on Ruby though because they
use a set of custom hooks that are implemented in it.

Refactor these scripts to use Bash instead of Ruby. This requires less
boilerplate code and gets rid of the last set of custom hooks in our
tests that use Ruby.