Makefile: Split up targets to build and install Git
We currently have a single target that's being used to both build and
install Git into the target location. While this target is used for
dependents to install a Git distribution via `make git GIT_PREFIX=...`,
it's also used as part of our tests to install into our own default Git
prefix so that we can set up the Git command factory correctly.

While the former use of this rule makes sense, it doesn't feel sensible
to require a complete installation for our tests. Let's prepare for a
fix to this by splitting up the rule into two: one to install a Git
distribution and one to build Git.