Makefile: Stop depending on an installed Git distribution for our tests
We don't need to have a complete Git distribution installed in order for
out tests to work. Instead, we can simply use Git's own bin-wrappers to
avoid installing the complete Git distribution so that we only have to
build the binaries.

Do so and convert our tests to only build the Git distribution without
installing it. This let's us get rid of the default Git installation we
always do when we execute tests and thus allows us to simplify the
target to install the Git distribution.