Makefile: Drop the default Git prefix
Now that our tests don't depend on the `git` target anymore but instead
us the bin-wrappers we can drop the default installation prefix of the
`git` target. Instead, we'll now set it up the same as the `PREFIX` by
default but keep it overridable. This both simplifies the logic we have
where we don't need to remove the default prefix anymore, but it also
causes us to pay proper attention to the `PREFIX` variable by default
when we install Git.

While the `GIT_PREFIX` variable is essentially useless now we still keep
it to remain compatible with existing calling conventions.