Makefile: Inline installation of Git distribution
Now that our tests don't depend on the `git` target anymore it should
really only ever be used to actually install a Git distribution. It's
thus not all that sensible anymore to have it depend on the target
location: Git's own Makefile should be the one to figure out what needs
updating and what doesn't.

Inline the recipe into the phony `git` rule.