Makefile: Introduce new `install-git` target
The `make git` target is kind of misnamed as it doesn't indicate that it
is about installing the Git distribution, but rather sounds like it
would only build it. Let's introduce a new `install-git` target that
clearly indicates that we're about to install it. This deprecates the
old `make git` rule, which is now simply an alias.