proto: Extend FindLicenseResponse with additional fields
Gitaly uses the Licensee gem to find the license used in a
repository and return the corresponding license identifier
to GitLab. GitLab uses the same Licensee gem to get additional
info about the license using the identifier to show it on
the UI. Go and Ruby implementations of the FindLicense RPC
rely on different algorithms for license detection and the
set of licenses that could be detected are different as well,
that is why they are not fully compatible. Let's extend
FindLicenseResponse with additional fields, so GitLab
doesn't need to get that info from the Licensee gem anymore.
It solves the problem of the compatibility of the two
implementations.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3782