golangci-lint: Document why we don't enable `t.Helper()` rule
The `thelper` linter has a rule that will flag test helper functions
that don't call `t.Helper()`. Unfortunately, this will also flag all
kinds of functions that really shouldn't have this call, which makes the
linter effectively useless for us.

Document why we disable it to not keep folks wondering.