objectpool: Default-enable pruning of refs to fix reference conflicts
In 018958fb1 (objectpool: Fix conflicting references when fetching into
pools, 2022-07-27) we have introduced the logic to enable pruning of
references when we update an object pool via FetchIntoObjectPool. This
change was required because we have seen cases where updating object
pools continued to fail due to conflicting references which exist in the
pool repository and the member we have been fetching from. By pruning
any references which don't exist in the member anymore we were able to
resolve these conflicts and thus this bug.

This change has been rolled out to production on August 3rd without any
observed issues. So let's default-enable the feature flag guarding this
code so that the fix will be included in the next release.

Changelog: fixed