license: Return additional info from FindLicense RPC
To remove the Licensee gem and replace it with the
go-license-detector package, in a previous commit, we extended
the interface of the FindLicense RPC to return additional
license details. Because implementations are different they
produce a bit different results for the same input. Let's try
to make it as much consistent as possible by adjusting
the license identifier (to lower-case) and the nickname
(we use a hardcoded dictionary of license id to nickname).

Because the changes we contributed to the
github.com/go-enry/go-license-detector package
are not yet applied by the maintainers, let's use a fork of it
with the changes we need. Once they are merged, we should update
the dependency.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3782