archive: Explicitly ignore error retruned by the tar writer
While we check error codes returned by the tar writer in the general
case when writing an archive, we ignore it in case we have already seen
an error in `TarBuilder.Close()`. We didn't explicitly ignore the error
with a blanket-assignment though, and thus the errcheck linter
complains.

Fix this by assigning the error to the blanket variable and drop the
corresponding exclude from our linting rules.