linguist: Ignore error in deferred close of zlib writer
We already check the error when we manually close the zlib writer, but
don't in the deferred call. This is fine though at it won't do anything
in the happy path as the writer was closed already, and we already got
an error in the unhappy path.

Explicitly ignore the error to make our linters happy.