gitlab: Copy client code from 'gitlab-shell'
Currently 'gitlab-shell' contains the code for the gitlab client. We use
this in our 'gitaly' code, but this causes a cyclic dependency since
'gitlab-shell' also imports parts of our code.

To remove this cyclic dependency, we decided to move most of the code
in-house. In this commit we do this by copying over the following files:
1. gitlabnet.go
2. httpclient.go
from 'gitlab-shell' without modification to a new 'client' package under
'internal/gitlab'.

We do not copy over the tests, since our repo has its own helpers for
tests which we'll leverage to rewrite the tests.