Include default labels in the default issue template
This commit adds commented out type:: labels in the default issue
template to make it easier to include the correct label. Additionally,
the default includes the relevant tags for categorizing Gitaly work
correctly. Empty lines are included in the beginning of the template so
one can just start typing out the issue without first having to create
new lines.