proto: Add UserCreateBranchError
Add a new message type `UserCreateBranchError`, which will be used to
provide structured errors for the `UserCreateBranch` RPC. This will be
released behind a feature flag.

Currently this structure only holds a single data type within it, i.e.
`CustomHookError`. Potentially can be extended in the future.

Run `make proto` to generate the related proto files from this.

Signed-off-by: Karthik Nayak <knayak@gitlab.com>