config: Remove DB metrics configuration
Scraping DB metrics requires expensive queries to the Praefect DB. By
default these metrics are split out to its own independent endpoint for
scraping. This reduces resouce consumption by enabling separate
configuration of the metrics scraping intervals. Currently the
`prometheus_exclude_database_from_default_metrics` configuration exists
to toggle this behavior. It is now desired for the default behavior to
be the only option. This change removes the configuration option and
makes it so DB metrics cannot be combined with other metrics on the same
endpoint.