featureflag: Add new flag 'user_create_branch_structured_errors'
Add a new flag to control the release of 'UserCreateBranchError' which
was introduced in the previous
commit (2d7e0ee3598caa9b127864c6bd8956ae80a9b8a9).

This flag will be used in the upcoming commits to enable the use of this
struct and eventually be deprecated once the struct is released and
adopted.