server factory: Wait for termination of all goroutines
It happens that goroutine started to Serve() doesn't terminate
after test completion and that fails the test (all started
goroutines must be terminated on test completion).
Now the test waits for all started goroutines to be terminated.
The result returned from the Server() method is also controlled
now and checked for correctness.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3916