proto: Return 'Branch' in 'FindLocalBranchesResponse'
Currently 'FindLocalBranchesResponse' contains 'FindLocalBranchResponse'
which is a custom type drawn over the common 'Branch' type. In due of
deprecating the old and keeping type consistency we introduce
'local_branches' as a new field of type 'Branch'.

In subsequent releases we will deprecate the old
'FindLocalBranchResponse' from 'FindLocalBranchesResponse' and only keep
'Branch' type in it.