repository: Add log only for new repository calculation methods
The RepositorySize is a high risk API, since an inaccurate number
(especially if it's much higher) can cause projects to be locked due to
usage quotas. In order to safely roll out a new calculation, add a log
only mode such that a feature flag has to be enabled in order to send
the value back to the Rails client. Without this feature flag, we merely
log the new calculation and continue to send the value calculated with
du.