blob: Enhance returned error messages
In some cases error returned by RPC methods are prefixed with
its names. This is redundant as the caller knows exactly what
method is called and also the log entry always contains full
method name including service name. Instead, the prefixes for
errors in some cases were added to better describe what operation
failed. It should help to better navigate to the reason of the
error.