blob: Replace %v with %w
We should use %w for the error wrapping. That will
allow us to unwrap the error without loosing any context.
It is not use right now, but maybe in the future if we need
to change response based on the returned error in the chain
of interceptors.
And one simplification from fmt.Errorf() to errors.New().