linguist: Use the colors from go-enry
We're about to go all-in on go-enry for the language detection. Which
means we'll soon no longer can load the language colors from a json file
provided by the linguist gem. We could copy the file, but go-enry
provides colors as well. So use those instead.

This also made it possible so turn the Color() function into a normal
function, instead of a receiver function.

This has a little disadvantage: it's no longer possible to load a custom
set of colors. We believe this isn't used, so we're willing to give up
on that.