hook: Export an error errInvalidSidechannelAddress
The errInvalidSidechannelAddress made exportable, so we can
check what is the cause of the error returned by the GetSidechannel()
function. It will be used to return InvalidArgument for RPC.
The error is also a bit refactored to have a value receiver for the
Error() method and now has explicit address field.