Makefile: Remove bundled git v2.35.0
In the commit (cc04215eb) we removed the flag to enable git v2.37.0.
Making it now the default git version. Now we can remove the older git
version v2.35.0. In this commit remove it from the Makefile. This means
it will no longer be bundled with Gitaly.

Also remove the patches added for git v2.35.0, which are no longer
required.