doc: Clarify GitLab/Gitaly boundary for RPCs
In the past it wasn't always clear where exactly we put the boundary for
RPCs that should be implemented in Gitaly versus those that should be in
Rails. This was recently clarified though: Gitaly should always avoid to
implement business-specific logic and policy. Instead, its provided RPCs
should focus on the problem-domains specific to Gitaly and Praefect.

Clarify this in our documentation to make it more readily discoverable
by engineers.