operations: Structured errors in UserCreateBranch
With 2e642ccd (branches: Use 'UserCreateBranchError' behind a flag) we
have started using structured errors for the UserCreateBranch RPC
instead of relying on `Internal` errors.

This code was rolled out to production and the feature flag is currently
enabled by default. Let's remove the feature flag guarding it.

Changelog: changed