praefect: run track-repositories tests serially
The `praefect` command has a global logger that is passed to any
subcommands executed. When executing normally this is fine as each
process will only run a single subcommand. However, in testing this
creates a risk of race conditions when tests are run in parallel, as
each test goroutine will share this logger.

To resolve race detector failures on `TestAddRepositories_Exec()` and
`TestAddRepositories_Exec()`, we will now run these tests serially. This
has minimal impact on test execution time, increasing from 4.663s to
5.129s.

In the medium-term we should replace the global logger with a local that
allows parallel testing. Issue #4500[0] has been opened for this.

[0] https://gitlab.com/gitlab-org/gitaly/-/issues/4500