git: Add `GetURLAndCurloptResolveConfig()`
To facilitate the creation of the ConfigPair for the
`http.curloptResolve` flag or modification of the URL for avoiding DNS
rebinding we introduce the `GetURLAndCurloptResolveConfig()` function.

This function takes in the remoteURL and the resolvedAddress and then
provides the modified URL and configPair to be passed onto Git.

For HTTP/HTTPS URLs we leave the remoteURL as is and use the
`http.curloptResolve` flag which was introduced in Git. This is done by
providing a new ConfigPair with the value in the format of
`HOST:PORT:ADDRESS[,ADDRESS]`.

For SSH/Git protocol URLs we simply replace the hostname with the
provided resolved address.

The function does basic validation of the provided remoteURL and
resolved address and throws errors if any validation failed.