config: Simplify error message
If relative path is not provided the error message used to include
the full Repository entity. It has no much of usage as we don't
rely on this data. With it the error message was hard to read and
to create an expected results in the tests. It is simplified to
have only an error message without the full Repository entity.