Praefect: Consolidate quorum check logic
When a node voter is updated a check is performed to determine whether
quorum has already been achieved or if quorum is still attainable. A
similar check is also performed to determine whether the blocked voters
in the pending subtransaction can be unblocked. Future changes to the
node RPC error handling will require that subtransactions signal voters
to unblock once quorum becomes impossible. This change consolidates the
quorum check logic into a single function that can be reused by
`updateVoterState()` and `mustSignalVoters()`.