Praefect: Get pending node subtransactions
Currently the logic to retrieve the pending subtransaction in a
transaction for a node is embedded in `getOrCreateSubtransaction()`.
With future changes to node RPC error handling this same functionality
will also be needed to cancel a voter associated with a failed node.
This change separates this logic into its own function so it can be
reused.