ci: Don't move go cache outside of project dir
With Gosec phased out in favor of the much faster Semgrep-based SAST
scanner, the performance implications of scanning the go cache are
minimal. Jobs which moved the cache[0] those that left it in-place[1]
are all taking roughly 20 to 30 seconds. We are also not seeing
vulnerabilities in depencies being reported against Gitaly itself, which
Gosec had been doing[2].

Remove the before_script to move the cache out of tree from the project
source.

[0] https://gitlab.com/gitlab-org/gitaly/-/jobs/3138452280#L29
[1] https://gitlab.com/gitlab-org/gitaly/-/jobs/3138531794#L29
[2] https://gitlab.com/gitlab-org/gitaly/-/merge_requests/4341