ref: Revert fix for DeleteRefs that broke QA
With 80d62284b (ref: Return structured error when update error is not
ErrAlreadyLocked, 2022-09-28) we have fixed a case where we still
returned successfully on error even though the feature flag to use
structured errors was enabled. This fix caused breakage in our QA jobs
though:

    1st Try error in ./spec/lib/gitlab/git/repository_spec.rb:463:
    expected Gitlab::Git::Repository::GitError, got #<Gitlab::Git::CommandError: 13:unable to prepare: state update to "prepare" failed: EOF, stderr: "fatal: invalid ref format: refs\\heads\\fix\n".> with backtrace:
    downstream QA

Revert this change for now to unblock deployments.