git: Validate \\ in refname
In ae728915 (refs: Return structured errors for DeleteRefs 2022-05-12),
we started to validate ref format and return a structured error. The
existence of a backslash is not something we currently validate, and
instead we rely on the `prepare` step to catch this error. We should
catch this in the ref validation step instead.

Changelog: changed