ruby: Constrain Redis Gem to version '~> 4.7.1'
Rails is currently blocked on the version upgrade to Redis v5.0.0 and
newer, but the upcoming bump to gitlab-labkit Gem will cause us to pull
in this major Redis version upgrade.

Constrain the Redis Gem version to '~> 4.7.1' to not pull in the new
version until Rails has upgraded. This is the same version that Rails
currently uses.