Makefile: Fix use of renamed CMake options for libgit2
The recent upgrade to libgit2 v1.5.0 brought in renames of two CMake
variables:

    - `-DTHREADSAFE` is now called `-DUSE_THREADS`.

    - `-DBUILD_CLAR` is now walled `-DBUILD_TESTS`.

We failed to adapt accordingly and thus we're currently using the
respective default values. As threads are enabled by default anyway this
doesn't matter, but we now fail to disable tests and thus accidentally
require Python to compile libgit2.

Fix this by adjusting the variables.