Praefect: Cancel voters with no vote
Currently voters can only be canceled if they have cast a vote.
Voters that have note yet voted should also have the ability to be
canceled in the event of a node RPC failure. This change updates
`updateVoterState()` to be able to cancel a voter and accurately track
vote counts regardless of whether the voter has voted.