tools/golangci-lint: Fix Viper v1.13 breaking the Revive linter config
The upgrade to Viper v1.13 has introduced a change to lower-case all
config entries in arrays via 5247643 (Recurse into arrays when
converting keys to lowercase, 2022-06-23). This in turn breaks our
ability to pass the `allowTypesBefore` argument to the Revive linter as
it performs an exact match for the arguments.

Let's unblock the upgrade to golangci-lint v1.50.0 by excluding Viper
v1.13 for the time being.